Fix warnings

+++ include/mp/os.h
@@ -121,7 +121,7 @@ class MemoryMappedFile : public internal::MemoryMappedFileBase {
     fmt::ULongLong unsigned_size = file_size;
     if (unsigned_size > std::numeric_limits<std::size_t>::max())
       throw Error("file is too big");
-    map(file, unsigned_size);
+    map(file, static_cast<std::size_t>(unsigned_size));
   }
 };
